Update .gitignore
Adding this line to the gitignore so that I can create a folder for stuff that's 'in development' that I can work on without pushing it to you guys. It will let me keep some of the auth stuff without deleting it, but in a way that it can still reference the project for my testing.  Just trying to do some 'cleaning up' before we merge streams to make this a bit easier and keep our final product clean.